// pix-front-project/src/app/shared/model/cobranca-vencimento.ts
export class CobrancaVencimento {
  txId? : string;
  dtVencimento?: Date | string;
  validadeAposVencimento?: number ;
  logradouroDevedor?: string;
  nomeCidadeDev?: number;
  ufDev?: string;
  cepDev?: string;
  vlrMulta?: number;
  modalidadeMulta?: number;
  vlrJuros?: number;
  modalidadeJuros?:number;
  vlrAbatimento?: number;
  modalidadeAbatimento?: number;
  logRec?:string;
  cidRec?:string;
  ufRec?:string;
  cepRec?:string;
}